refactor(vspreview): split session script generation
